backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = "^0.104.1"
starlette = "^0.19.0"
uvicorn = "^0.17.6"
pydantic = "^1.9.0"
databases = "^0.5.5"
python-dotenv = "^0.20.0"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
mypy = "^0.942"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
